+++ src/main/java/in/oxidane/work/done/project/dto/ProjectStatusResponse.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class ProjectStatusResponse {
 
     @Schema(description = "Unique identifier for the project status", example = "1")
-    private Integer projectStatusId;
+    private Long projectStatusId;
 
     @Schema(description = "Name of the project status", example = "In Progress")
     private String projectStatusName;
