+++ src/main/java/in/oxidane/work/done/project/validator/ProjectScopeValidator.java
@@ -1,6 +1,6 @@
 package in.oxidane.work.done.project.validator;
 
-import in.oxidane.work.done.exception.ValidationException;
+import in.oxidane.work.done.common.exception.ValidationException;
 import in.oxidane.work.done.project.dto.ProjectScopeRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -37,7 +37,7 @@ public class ProjectScopeValidator {
      * @param id The ID of the project scope being updated
      * @throws ValidationException if validation fails
      */
-    public void validateForUpdate(ProjectScopeRequest request, Integer id) {
+    public void validateForUpdate(ProjectScopeRequest request, Long id) {
         log.debug("Validating project scope update request for id: {}", id);
         if (id == null) {
             throw new ValidationException("Project scope ID cannot be null for update");
